+++ src/interfaces/DivingCylinderSet.ts
@@ -1,0 +1,15 @@
+export type DivingCylinder = {
+  id: string;
+  volume: string;
+  material: string;
+  pressure: string;
+  serialNumber: string;
+  inspection: string;
+};
+
+export type DivingCylinderSet = {
+  id: string;
+  owner: string;
+  name: string;
+  cylinders: DivingCylinder[];
+};
