+++ src/lib/queries/divingCylinderQuery.ts
@@ -1,9 +1,9 @@
 import { useQuery } from '@tanstack/react-query';
 import { toast } from 'react-toastify';
 import {
-  DivingCylinderSet,
   getDivingCylinderSets,
 } from '../apiRequests/divingCylinderSetRequests';
+import { DivingCylinderSet } from '../../interfaces/DivingCylinderSet';
 import { UseQuery } from './common';
 import { CYLINDER_SETS_QUERY_KEY } from './queryKeys';
